+++ Frontend/src/components/Molecules/ButtonWithLabel/ButtonWithLabel.js
@@ -1,10 +1,15 @@
 import React from 'react';
 import Paragraph from '../../Atoms/Typography/Paragraph/Paragraph';
 import DivImage from '../../Atoms/DivImage/DivImage';
+import './ButtonWithLabel.scss';
 
-const ButtonWithLabel = ({children, path}) =>{
+const ButtonWithLabel = ({children, onClickFn = ()=>{}, path}) =>{
+    const onClickHandler = (e) => {
+        onClickFn();
+    }
+
     return(
-        <div>
+        <div className="ButtonWithLabel" onClick={onClickHandler}>
             <DivImage path={path}/>
             <Paragraph>{children}</Paragraph>
         </div>