+++ Frontend/src/views/BasicView.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import MainTemplate from '../templates/MainTemplate';
+
+const BasicView = () => {
+    return <MainTemplate />
+}
+
+export default BasicView;