+++ Frontend/src/views/NewSetView.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import MainTemplate from '../templates/MainTemplate';
+
+const NewSetView = () => {
+    return (
+    <MainTemplate>
+        <p>NewSetView</p>
+    </MainTemplate>
+    );
+};
+
+export default NewSetView;