+++ changelog.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2020.06.07
+
+**Added**
+- Dependencies : React-router-dom,
+- Images : new.svg,
+- Views : BasicView, NewSetView.
+- Compontents: DivImage, ButtonWithLabel, Paragraph, MainTemplate.
+
+**Changes**
+- Sidebar can navigate between pages.
+- Changed extension of Colors.js to .scss.
+
 ## [0.0.1] - 2020.06.07
 
 **Added**